Resizing figures has been added. And a little error fixed when saving.
->If you want to save while a figure is selected, a runtime error will occur. So, Before saving I had to call pManager->removeSelection();

->This error is because the draw color of the selected figure is magenta. And we don't want to save magenta. We want to save the original draw color. (SaveAction.cpp throws an exception when the color isn't red, green, blue, black or white)

+++ Phase2 - Project Framework/Actions/SaveLoad/LoadAction.h
@@ -15,7 +15,7 @@ class LoadAction : public Action
 	FileLinesFormat getFigureFromLine(float firstWordInLine);
 	void setLineDescription(FileLinesFormat lineFormat, const WordType*& descriptionArr, int& arrSize);
 	bool getCorrectWordType(int wordCounter, const WordType* descriptionArr, int size, WordType& type);
-	void getWordRangeExclusive(WordType type, int& min, int& max);
+	void getWordRangeExclusive(WordType type, float & min, float & max);
 	bool validateWord(float currentWord, WordType correctType);
 
 	void translateFigureWords(const float* figureWords, CFigure** loadedFigures, int& currentLoadedFiguresIndex);
